Simplify InputQueue implementation, put it into its own file
