+++ CMakeLists.txt
@@ -207,6 +207,7 @@ add_library(${TARGET_CORE_NAME}
     src/pipeline/AssetManager.cpp
     src/pipeline/MessageQueue.cpp
     src/pipeline/Node.cpp
+    src/pipeline/InputQueue.cpp
     src/pipeline/ThreadedNode.cpp
     src/pipeline/DeviceNode.cpp
     src/pipeline/node/XLinkIn.cpp
